third_party/WebKit/LayoutTests/accessibility/line-for-index-endless-loop.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<style>
.hideAllContainers .container {
    display: none;
}
</style>

<div class="container" id="contentEditable" contentEditable>
  <span style="display: inline-block;">
    <span style="position: absolute;">
      <span>x</span>
    </span>
    <span>2</span>
    <span style="display: inline-block;"></span>
  </span>
  <span>+</span>
</div>

<script>
test(function(t) {
    var axContentEditable = accessibilityController.accessibleElementById("contentEditable");
    assert_equals(axContentEditable.lineForIndex(0), 0);
}, "No endless loop when getting line for index in contentEditable");
</script>

<script>
if (window.testRunner)
    document.body.className = "hideAllContainers";
</script>
